HR Management System
Demo project created in which HR can access, check or update the attendance, leaverequest, salary, department of employees and can also see the employees name using employee table. In frontend all the data are static except leaverequest, React JS is used in Frontend part. In backend ASP.NET CORE is used and SQL as a database.